pyproject.toml
[project]
name = "feature-visualization"
version = "0.1.0"
description = "Scripts for visualizing feature extraction process"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.8",
    "matplotlib>=3.10.0",
    "pandas>=2.2.3",
    "pyradiomics",
]

[[tool.uv.index]]
url = "https://mirrors.ustc.edu.cn/pypi/simple"
default = true

[tool.uv.sources]
pyradiomics = { git = "https://github.com/bioinformatist/pyradiomics.git" }
